requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.7
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
python-dotenv==1.0.0
Werkzeug==2.3.7
SQLAlchemy==2.0.21
alembic==1.12.0 